# apps/api/openapi.yaml
openapi: 3.0.3
info:
  title: FeedbackHub API
  version: "1.0.0"
  description: |
    REST API for FeedbackHub, a self-hostable feedback & survey widget platform.
    Administrative endpoints require a Bearer access token. Widget-facing endpoints
    (under /api/v1/public and POST /api/v1/responses) are authenticated with a
    project public key instead of a user session.
servers:
  - url: http://localhost:4000/api/v1
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    projectKey:
      type: apiKey
      in: header
      name: X-Project-Key
...
paths:
  /auth/register:
    post:
      summary: Create an account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
                name: { type: string }
      responses:
        "201": { description: Account created }
  /auth/login:
    post:
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        "200": { description: Access token issued, refresh token set as HTTP-only cookie }
...
  /auth/me:
    get:
      summary: Get the current authenticated user
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Current user }
  /organisations:
    get:
      summary: List organisations the current user belongs to
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: List of organisations }
    post:
      summary: Create an organisation (creator becomes OWNER)
      security: [{ bearerAuth: [] }]
      responses:
        "201": { description: Organisation created }
...
  /api-keys:
    get:
      summary: List API keys for a project
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: projectId, in: query, required: true, schema: { type: string } }
      responses:
        "200": { description: List of API keys (secret values redacted) }
    post:
      summary: Create an API key (public or secret). Secret is returned once.
      security: [{ bearerAuth: [] }]
      responses:
        "201": { description: API key created }
  /api-keys/{id}/rotate:
    post:
      summary: Revoke and reissue an API key
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: New API key }
...